added new icon library and move style sheets
