+++ index.html
@@ -3,8 +3,10 @@
 <head>
   <meta charset="utf-8">
   <title>Weatherly</title>
-  <link href="normalize.css" type="text/css" rel="stylesheet">
+  <link href="./styles/normalize.css" type="text/css" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css?family=Questrial" rel="stylesheet">
+  <link rel="stylesheet" href="./styles/weather-icons.min.css">
+  <link rel="stylesheet" href="./styles/weather-icons-wind.min.css">
 </head>
 <body>
   <div id="app"></div>
