+++ lib/App.js
@@ -6,7 +6,7 @@ import Search from './Search';
 import APIKey from './.APIKey';
 import cleanData from './cleanData';
 import TenDay from './TenDay';
-import './app.css';
+import '../styles/app.css';
 
 
 export default class App extends React.Component {
@@ -38,7 +38,8 @@ export default class App extends React.Component {
 	}
 
 	componentDidMount() {
-		this.fetchData(this.state.state, this.state.city)
+		this.fetchData(this.state.state, this.state.city);
+		console.log('component mounted');
 	}
 
 	render() {
