+++ lib/Card.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './card.css';
+import '../styles/card.css';
 
 const Card = (props) => {
 	return (
@@ -9,7 +9,7 @@ const Card = (props) => {
 				<h1>{props.time}</h1>
 				<h2>{props.temp}</h2>
 				<h3>{props.cond}</h3>
-
+				<i className="wi wi-night-sleet"></i>
 			</div>
 			<div className="day-cards">
 				<h3>{props.day}</h3>
