+++ lib/CurrentWeather.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Card from './Card';
-import './currentWeather.css';
+import '../styles/currentWeather.css';
 
 const CurrentWeather = (props) => {
 	return (
