+++ lib/Search.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './search.css';
+import '../styles/search.css';
 
 class Search extends React.Component {
 	constructor(props) {
