+++ lib/SevenHour.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Card from './Card';
-import './sevenHour.css';
+import '../styles/sevenHour.css';
+import FontAwesome from 'react-fontawesome';
 
 const SevenHour = (props) => {
 	return (
@@ -9,6 +10,7 @@ const SevenHour = (props) => {
 			{
         props.sevenHour.map( (card, index) => {
           return (
+            
             <Card className="hour-card"
                   time={card.time} 
             			temp={card.temp}
