+++ lib/TenDay.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Card from './Card';
-import './tenDay.scss';
+import '../styles/tenDay.scss';
 
 const TenDay = (props) => {
   return (
