+++ lib/Welcome.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Search from './Search';
-import './welcome.scss';
+import '../styles/welcome.scss';
 
 const Welcome = () => {
   return (
