+++ lib/cleanData.js
@@ -1,4 +1,5 @@
 function cleanData(data) {
+  console.log(data);
  
  	////////////  CURRENT WEATHER  /////////////
 	let timeDay = data.current_observation.observation_time.split('');
